+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## `0.2.2`
+
+- Fixed `hsv2rgb()` array hue going infinity because of in-place array mutation
+
+## `0.2.1`
+
 - `hsv2rgb()` can now take a single argument for grayscale value (0..1)
 
 ## `0.2.0`